Allow trip owner to add other users to trip to either be editor or viewer.

+++ app/src/main/java/sg/edu/np/mad/assignment/TripAdapter.java
@@ -26,11 +26,14 @@ import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.dynamiclinks.DynamicLink;
 import com.google.firebase.dynamiclinks.FirebaseDynamicLinks;
-import com.google.firebase.dynamiclinks.ShortDynamicLink;
+import com.google.firebase.firestore.DocumentSnapshot;
 import com.google.firebase.firestore.FirebaseFirestore;
+import com.google.firebase.firestore.QuerySnapshot;
+import com.google.gson.Gson;
 
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
+import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.Date;
 import java.util.List;
@@ -73,18 +76,11 @@ public class TripAdapter extends RecyclerView.Adapter<TripAdapter.myviewholder>
         String durationStr = dataHolder.get(position).getStartDate() + " - " + dataHolder.get(position).getEndDate();
         holder.duration.setText(durationStr);
 
-        //Omit slashes in date
-        String dateNoSlash = holder.duration.getText().toString();
-
-        if (dateNoSlash.contains("/")){
-            dateNoSlash = dateNoSlash.replace("/", " ");
-            holder.duration.setText(dateNoSlash);
-        }
 
         // Set time left till start of trip
         // get current date
         calendar = Calendar.getInstance();
-        dateFormat = new SimpleDateFormat("dd/MMM/yyyy");
+        dateFormat = new SimpleDateFormat("d MMM yyyy");
         todaydate = dateFormat.format(calendar.getTime());
 
         try {
@@ -127,23 +123,80 @@ public class TripAdapter extends RecyclerView.Adapter<TripAdapter.myviewholder>
                 MenuInflater inflater = popupMenu.getMenuInflater();
                 inflater.inflate(R.menu.popup_menu, popupMenu.getMenu());
 
-
                 // Set collab menu item to invisible
                 MenuItem collabItem = popupMenu.getMenu().findItem(R.id.collabMenu);
                 collabItem.setVisible(false);
 
-                // Set edit menu to invisible if user got no permission to edit
                 MenuItem editItem = popupMenu.getMenu().findItem(R.id.editMenu);
+                MenuItem deleteItem = popupMenu.getMenu().findItem(R.id.delMenu);
 
-                if (uid.equals(trip.getUserId()))
+                // If is not owner of trip
+                if (!uid.equals(trip.getUserId()))
                 {
-                    editItem.setVisible(true);
-                }
-                else
-                {
-                    editItem.setVisible(false);
+                    // user cant delete trip
+                    deleteItem.setVisible(false);
 
                 }
+
+                db.collection("Trip").get().addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
+                    @Override
+                        public void onComplete(@NonNull Task<QuerySnapshot> task) {
+                        for (DocumentSnapshot doc : task.getResult()) {
+                          if (trip.getId().equals(doc.getString("id"))) {
+                              if (task.isSuccessful()) {
+                                  ArrayList<String> sharedTripLists = new ArrayList<String>();
+                                  ArrayList<String> stalist = new ArrayList<String>();
+                                  stalist = (ArrayList<String>) doc.get("serializedTAL");
+
+                                  for (int i=0; i<stalist.size(); i++){
+                                      Gson gson = new Gson();
+                                      TripAdmin tempTa = gson.fromJson(stalist.get(i), TripAdmin.class);
+//                                      sharedTripLists.add(tempTa.userId);
+
+                                      if (uid.equals(doc.getString("userId")) || (tempTa.getUserId().equals(uid) && tempTa.permission.equals("Can Edit")))
+                                      {
+                                          editItem.setVisible(true);
+                                      }
+                                      else
+                                      {
+                                          // cant edit trip
+                                          editItem.setVisible(false);
+                                      }
+                                  }
+
+                              }
+                          }
+                        }
+                    }
+                });
+
+//                db.collection("Trip").document(trip.getId()).get().addOnCompleteListener(new OnCompleteListener<DocumentSnapshot>() {
+//                    @Override
+//                    public void onComplete(@NonNull Task<DocumentSnapshot> task) {
+//                        if (task.isSuccessful()) {
+//                            ArrayList<String> sharedTripLists = new ArrayList<String>();
+//                            ArrayList<String> stalist = new ArrayList<String>();
+//                            stalist = (ArrayList<String>) task.getResult().get("serializedTAL");
+//
+//                            for (int i=0; i<stalist.size(); i++){
+//                                Gson gson = new Gson();
+//                                TripAdmin tempTa = gson.fromJson(stalist.get(i), TripAdmin.class);
+//                                sharedTripLists.add(tempTa.userId);
+//                            }
+//                            Log.v("stupid", String.valueOf(sharedTripLists.contains(uid)));
+//                            if (uid.equals(task.getResult().getString("userId")) || sharedTripLists.contains(uid))
+//                            {
+//                                editItem.setVisible(true);
+//                            }
+//                            else
+//                            {
+//                                // cant edit trip
+//                                editItem.setVisible(false);
+//                            }
+//                        }
+//
+//                    }
+//                });
 
                 // When selecting item menu
                 popupMenu.setOnMenuItemClickListener(new PopupMenu.OnMenuItemClickListener() {
@@ -201,34 +254,42 @@ public class TripAdapter extends RecyclerView.Adapter<TripAdapter.myviewholder>
 
                             case R.id.delMenu:
                                 new AlertDialog.Builder(view.getContext())
-                                        .setMessage("Delete this trip?")
-                                        .setPositiveButton("DELETE", new DialogInterface.OnClickListener()
-                                        {
-                                            @Override
-                                            public void onClick(DialogInterface dialog, int which) {
-                                                db.collection("Trip").document(trip.getId())
-                                                        .delete()
-                                                        .addOnSuccessListener(new OnSuccessListener<Void>() {
-                                                            @Override
-                                                            public void onSuccess(Void aVoid) {
-                                                                dataHolder.remove(position);
-                                                                notifyItemRemoved(position);
-                                                                Log.d("DeleteTrip", "DocumentSnapshot successfully deleted!");
-                                                            }
-                                                        })
-                                                        .addOnFailureListener(new OnFailureListener() {
-                                                            @Override
-                                                            public void onFailure( Exception e) {
-                                                                Log.w("DeleteTrip", "Error deleting document", e);
-                                                            }
-                                                        });
-                                            }
+                                    .setMessage("Delete this trip?")
+                                    .setPositiveButton("DELETE", new DialogInterface.OnClickListener()
+                                    {
+                                        @Override
+                                        public void onClick(DialogInterface dialog, int which) {
+                                            db.collection("Trip").get().addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
+                                                @Override
+                                                public void onComplete(@NonNull Task<QuerySnapshot> task) {
+                                                    for (DocumentSnapshot doc : task.getResult()) {
+                                                        if (trip.getId().equals(doc.getString("id"))) {
+                                                            db.collection("Trip").document(doc.getId())
+                                                                .delete()
+                                                                .addOnSuccessListener(new OnSuccessListener<Void>() {
+                                                                    @Override
+                                                                    public void onSuccess(Void aVoid) {
+                                                                        dataHolder.remove(position);
+                                                                        notifyItemRemoved(position);
+                                                                        Log.d("DeleteTrip", "DocumentSnapshot successfully deleted!");
+                                                                    }
+                                                                })
+                                                                .addOnFailureListener(new OnFailureListener() {
+                                                                    @Override
+                                                                    public void onFailure( Exception e) {
+                                                                        Log.w("DeleteTrip", "Error deleting document", e);
+                                                                    }
+                                                                });
+                                                        }
+                                                    }
+                                                }
+                                            });
 
-                                        })
-                                        .setNegativeButton("Cancel", null)
-                                        .show();
+                                        }
 
-
+                                    })
+                                    .setNegativeButton("Cancel", null)
+                                    .show();
                                 break;
                         }
                         return true;
